dcgan

+++ DCGAN1.py
@@ -61,7 +61,7 @@ for epoch in range(num_epochs):
         obs = obs.to(device)
 
         # Train Discriminator
-        rand = torch.randn((obs[0].shape[0], z_dim, 1, 1), device=device, dtype=obs.dtype)
+        rand = torch.randn((obs[0].shape[0], z_dim, 1, 1), device=device, dtype=torch.float64)
         action_ = generator(rand)
         action = torch.cat([obs.view_as(action_), action_], 0)
         reward_ = torch.zeros((len(obs), 1)).to(obs)
